pyproject.toml
[project]
name = "pc-uart-proto"
version = "0.1.0"
description = "PC-side CLI program for UART communication using Protobuf"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
]
